EscapeAnalysis is a simple module that collects escape information in Julia's SSA optimization IR a.k.a. IRCode.

You can give a try to the escape analysis with the convenience entries that EscapeAnalysis exports for testing and debugging purposes:

EscapeAnalysis.EAUtils.code_escapes — Function

code_escapes(f, argtypes=Tuple{}; [world], [interp]) -> result::EscapeResult
code_escapes(tt::Type{<:Tuple}; [world], [interp]) -> result::EscapeResult

Runs the escape analysis on optimized IR of a generic function call with the given type signature. Note that the escape analysis runs after inlining, but before any other optimizations.

julia> mutable struct SafeRef{T}
           x::T
       end

julia> Base.getindex(x::SafeRef) = x.x;

julia> Base.isassigned(x::SafeRef) = true;

julia> get′(x) = isassigned(x) ? x[] : throw(x);

julia> result = code_escapes((String,String,String,String)) do s1, s2, s3, s4
           r1 = Ref(s1)
           r2 = Ref(s2)
           r3 = SafeRef(s3)
           try
               s1 = get′(r1)
               ret = sizeof(s1)
           catch err
               global g = err # will definitely escape `r1`
           end
           s2 = get′(r2)      # still `r2` doesn't escape fully
           s3 = get′(r3)      # still `r3` doesn't escape fully
           s4 = sizeof(s4)    # the argument `s4` doesn't escape here
           return s2, s3, s4
       end
#1(X _2::String, ↑ _3::String, ↑ _4::String, ✓ _5::String) in Main at REPL[6]:2
2  X  1 ── %1  = %new(Base.RefValue{String}, _2)::Base.RefValue{String}            │╻╷╷ Ref
3  *′ │    %2  = %new(Base.RefValue{String}, _3)::Base.RefValue{String}            │╻╷╷ Ref
4  ✓′ └─── %3  = %new(SafeRef{String}, _4)::SafeRef{String}                        │╻╷  SafeRef
5  ◌  2 ── %4  = $(Expr(:enter, #8))                                               │
   ✓′ │    %5  = ϒ (%3)::SafeRef{String}                                           │
   *′ │    %6  = ϒ (%2)::Base.RefValue{String}                                     │
   ✓  └─── %7  = ϒ (_5)::String                                                    │
6  ◌  3 ── %8  = Base.isdefined(%1, :x)::Bool                                      │╻╷  get′
   ◌  └───       goto #5 if not %8                                                 ││
   X  4 ──       Base.getfield(%1, :x)::String                                     ││╻   getindex
   ◌  └───       goto #6                                                           ││
   ◌  5 ──       Main.throw(%1)::Union{}                                           ││
   ◌  └───       unreachable                                                       ││
7  ◌  6 ──       nothing::typeof(Core.sizeof)                                      │╻   sizeof
   ◌  │          nothing::Int64                                                    ││
   ◌  └───       $(Expr(:leave, 1))                                                │
   ◌  7 ──       goto #10                                                          │
   ✓′ 8 ── %18 = φᶜ (%5)::SafeRef{String}                                          │
   *′ │    %19 = φᶜ (%6)::Base.RefValue{String}                                    │
   ✓  │    %20 = φᶜ (%7)::String                                                   │
   ◌  └───       $(Expr(:leave, 1))                                                │
   X  9 ── %22 = $(Expr(:the_exception))::Any                                      │
9  ◌  │          (Main.g = %22)::Any                                               │
   ◌  └───       $(Expr(:pop_exception, :(%4)))::Any                               │
11 ✓′ 10 ┄ %25 = φ (#7 => %3, #9 => %18)::SafeRef{String}                          │
   *′ │    %26 = φ (#7 => %2, #9 => %19)::Base.RefValue{String}                    │
   ✓  │    %27 = φ (#7 => _5, #9 => %20)::String                                   │
   ◌  │    %28 = Base.isdefined(%26, :x)::Bool                                     ││╻   isassigned
   ◌  └───       goto #12 if not %28                                               ││
   ↑  11 ─ %30 = Base.getfield(%26, :x)::String                                    │││╻   getproperty
   ◌  └───       goto #13                                                          ││
   ◌  12 ─       Main.throw(%26)::Union{}                                          ││
   ◌  └───       unreachable                                                       ││
12 ↑  13 ─ %34 = Base.getfield(%25, :x)::String                                    │╻╷╷ get′
13 ◌  │    %35 = Core.sizeof::typeof(Core.sizeof)                                  │╻   sizeof
   ◌  │    %36 = (%35)(%27)::Int64                                                 ││
14 ↑  │    %37 = Core.tuple(%30, %34, %36)::Tuple{String, String, Int64}           │
   ◌  └───       return %37                                                        │

The symbols in the side of each call argument and SSA statements represents the following meaning:

◌ (plain): this value is not analyzed because escape information of it won't be used anyway (when the object is isbitstype for example)

✓ (green or cyan): this value never escapes (has_no_escape(result.state[x]) holds), colored blue if it has arg escape also (has_arg_escape(result.state[x]) holds)

↑ (blue or yellow): this value can escape to the caller via return (has_return_escape(result.state[x]) holds), colored yellow if it has unhandled thrown escape also (has_thrown_escape(result.state[x]) holds)

X (red): this value can escape to somewhere the escape analysis can't reason about like escapes to a global memory (has_all_escape(result.state[x]) holds)

* (bold): this value's escape state is between the ReturnEscape and AllEscape in the partial order of EscapeInfo, colored yellow if it has unhandled thrown escape also (has_thrown_escape(result.state[x]) holds)

′: this value has additional object field / array element information in its AliasInfo property

For testing, escape information of each call argument and SSA value can be inspected programmatically as like:

julia> result.state[Core.Argument(3)]
ReturnEscape

julia> result.state[Core.SSAValue(3)]
NoEscape′

EscapeAnalysis.EAUtils.@code_escapes — Macro

@code_escapes [options...] f(args...)

Evaluates the arguments to the function call, determines its types, and then calls code_escapes on the resulting expression. As with @code_typed and its family, any of code_escapes keyword arguments can be given as the optional arguments like @code_escapes interp=myinterp myfunc(myargs...).

Analysis Design

Lattice Design

EscapeAnalysis is implemented as a data-flow analysis that works on a lattice of x::EscapeInfo, which is composed of the following properties:

x.Analyzed::Bool: not formally part of the lattice, only indicates x has not been analyzed or not

x.ReturnEscape::BitSet: records SSA statements where x can escape to the caller via return

x.ThrownEscape::BitSet: records SSA statements where x can be thrown as exception (used for the exception handling described below)

x.AliasInfo: maintains all possible values that can be aliased to fields or array elements of x (used for the alias analysis described below)

x.ArgEscape::Int (not implemented yet): indicates it will escape to the caller through setfield! on argument(s)

These attributes can be combined to create a partial lattice that has a finite height, given the invariant that an input program has a finite number of statements, which is assured by Julia's semantics. The clever part of this lattice design is that it enables a simpler implementation of lattice operations by allowing them to handle each lattice property separately[LatticeDesign].

Backward Escape Propagation

This escape analysis implementation is based on the data-flow algorithm described in the paper[MM02]. The analysis works on the lattice of EscapeInfo and transitions lattice elements from the bottom to the top until every lattice element gets converged to a fixed point by maintaining a (conceptual) working set that contains program counters corresponding to remaining SSA statements to be analyzed. The analysis manages a single global state that tracks EscapeInfo of each argument and SSA statement, but also note that some flow-sensitivity is encoded as program counters recorded in EscapeInfo's ReturnEscape property, which can be combined with domination analysis later to reason about flow-sensitivity if necessary.

One distinctive design of this escape analysis is that it is fully backward, i.e. escape information flows from usages to definitions. For example, in the code snippet below, EA first analyzes the statement return %1 and imposes ReturnEscape on %1 (corresponding to obj), and then it analyzes %1 = %new(Base.RefValue{String, _2})) and propagates the ReturnEscape imposed on %1 to the call argument _2 (corresponding to s):

julia> code_escapes((String,)) do s
           obj = Ref(s)
           return obj
       end
#1(↑ _2::String) in Main at REPL[2]:2
2 ↑  1 ─ %1 = %new(Base.RefValue{String}, _2)::Base.RefValue{String}                │╻╷╷ Ref
3 ◌  └──      return %1                                                             │

The key observation here is that this backward analysis allows escape information to flow naturally along the use-def chain rather than control-flow[BackandForth]. As a result this scheme enables a simple implementation of escape analysis, e.g. PhiNode for example can be handled simply by propagating escape information imposed on a PhiNode to its predecessor values:

julia> code_escapes((Bool, String, String)) do cnd, s, t
           if cnd
               obj = Ref(s)
           else
               obj = Ref(t)
           end
           return obj
       end
  #3(↑ _2::Bool, ↑ _3::String, ↑ _4::String) in Main at REPL[3]:2
2 ◌  1 ─      goto #3 if not _2                                                     │
3 ↑  2 ─ %2 = %new(Base.RefValue{String}, _3)::Base.RefValue{String}                │╻╷╷ Ref
  ◌  └──      goto #4                                                               │
5 ↑  3 ─ %4 = %new(Base.RefValue{String}, _4)::Base.RefValue{String}                │╻╷╷ Ref
7 ↑  4 ┄ %5 = φ (#2 => %2, #3 => %4)::Base.RefValue{String}                         │
  ◌  └──      return %5                                                             │

Alias Analysis

EscapeAnalysis implements a backward field analysis in order to reason about escapes imposed on object fields with certain accuracy, and x::EscapeInfo's x.AliasInfo property exists for this purpose. It records all possible values that can be aliased to fields of x at "usage" sites, and then the escape information of that recorded values are propagated to the actual field values later at "definition" sites. More specifically, the analysis records a value that may be aliased to a field of object by analyzing getfield call, and then it propagates its escape information to the field when analyzing %new(...) expression or setfield! call[Dynamism].

julia> mutable struct SafeRef{T}
           x::T
       end

julia> Base.getindex(x::SafeRef) = x.x;

julia> Base.setindex!(x::SafeRef, v) = x.x = v;

julia> code_escapes((String,)) do s
           obj = SafeRef("init")
           obj[] = s
           v = obj[]
           return v
       end
#5(↑ _2::String) in Main at REPL[7]:2
2 ✓′ 1 ─ %1 = %new(SafeRef{String}, "init")::SafeRef{String}                   │╻╷ SafeRef
3 ◌  │        Base.setfield!(%1, :x, _2)::String                               │╻╷ setindex!
4 ↑  │   %3 = Base.getfield(%1, :x)::String                                    │╻╷ getindex
5 ◌  └──      return %3                                                        │

In the example above, ReturnEscape imposed on %3 (corresponding to v) is not directly propagated to %1 (corresponding to obj) but rather that ReturnEscape is only propagated to _2 (corresponding to s). Here %3 is recorded in %1's AliasInfo property as it can be aliased to the first field of %1, and then when analyzing Base.setfield!(%1, :x, _2)::String, that escape information is propagated to _2 but not to %1.

So EscapeAnalysis tracks which IR elements can be aliased across a getfield-%new/setfield! chain in order to analyze escapes of object fields, but actually this alias analysis needs to be generalized to handle other IR elements as well. This is because in Julia IR the same object is sometimes represented by different IR elements and so we should make sure that those different IR elements that actually can represent the same object share the same escape information. IR elements that return the same object as their operand(s), such as PiNode and typeassert, can cause that IR-level aliasing and thus requires escape information imposed on any of such aliased values to be shared between them. More interestingly, it is also needed for correctly reasoning about mutations on PhiNode. Let's consider the following example:

julia> code_escapes((Bool, String,)) do cond, x
           if cond
               ϕ2 = ϕ1 = SafeRef("foo")
           else
               ϕ2 = ϕ1 = SafeRef("bar")
           end
           ϕ2[] = x
           y = ϕ1[]
           return y
       end
#7(↑ _2::Bool, ↑ _3::String) in Main at REPL[8]:2
2 ◌  1 ─      goto #3 if not _2                                                │
3 ✓′ 2 ─ %2 = %new(SafeRef{String}, "foo")::SafeRef{String}                    │╻╷ SafeRef
  ◌  └──      goto #4                                                          │
5 ✓′ 3 ─ %4 = %new(SafeRef{String}, "bar")::SafeRef{String}                    │╻╷ SafeRef
7 ✓′ 4 ┄ %5 = φ (#2 => %2, #3 => %4)::SafeRef{String}                          │
  ✓′ │   %6 = φ (#2 => %2, #3 => %4)::SafeRef{String}                          │
  ◌  │        Base.setfield!(%5, :x, _3)::String                               │╻  setindex!
8 ↑  │   %8 = Base.getfield(%6, :x)::String                                    │╻╷ getindex
9 ◌  └──      return %8                                                        │

ϕ1 = %5 and ϕ2 = %6 are aliased and thus ReturnEscape imposed on %8 = Base.getfield(%6, :x)::String (corresponding to y = ϕ1[]) needs to be propagated to Base.setfield!(%5, :x, _3)::String (corresponding to ϕ2[] = x). In order for such escape information to be propagated correctly, the analysis should recognize that the predecessors of ϕ1 and ϕ2 can be aliased as well and equalize their escape information.

One interesting property of such aliasing information is that it is not known at "usage" site but can only be derived at "definition" site (as aliasing is conceptually equivalent to assignment), and thus it doesn't naturally fit in a backward analysis. In order to efficiently propagate escape information between related values, EscapeAnalysis.jl uses an approach inspired by the escape analysis algorithm explained in an old JVM paper[JVM05]. That is, in addition to managing escape lattice elements, the analysis also maintains an "equi"-alias set, a disjoint set of aliased arguments and SSA statements. The alias set manages values that can be aliased to each other and allows escape information imposed on any of such aliased values to be equalized between them.

Lastly, the alias/field analysis described above can also be generalized to analyze array operations. There is one big difference that object fields can be analyzed precisely using type information derived by inference while inference currently can't provide array dimensions that are necessary for analyzing array indexing precisely. Because of this reason EscapeAnalysis at this moment first does an additional simple linear scan to analyze dimensions of allocated arrays so that the succeeding escape analysis can analyze operations on those arrays with precise index information later. Still the simple linear scan can't reason about complicated aliasing, e.g. aliases via nested arrayset calls, and in such cases EscapeAnalysis gives up precise indexing analysis and switches to the more imprecise analysis that doesn't track precise index information.

Exception Handling

It would be also worth noting how EscapeAnalysis handles possible escapes via exceptions. Naively it seems enough to propagate escape information imposed on :the_exception object to all values that may be thrown in a corresponding try block. But there are actually several other ways to access to the exception object in Julia, such as Base.current_exceptions and manual catch of rethrown object. For example, escape analysis needs to account for potential escape of r in the example below:

julia> const Gx = Ref{Any}();

julia> @noinline function rethrow_escape!()
           try
               rethrow()
           catch err
               Gx[] = err
           end
       end;

julia> get′(x) = isassigned(x) ? x[] : throw(x);

julia> code_escapes() do
           r = Ref{String}()
           local t
           try
               t = get′(r)
           catch err
               t = typeof(err)   # `err` (which `r` aliases to) doesn't escape here
               rethrow_escape!() # but `r` escapes here
           end
           return t
       end
#9() in Main at REPL[12]:2
2  X  1 ── %1  = %new(Base.RefValue{String})::Base.RefValue{String}            │╻╷ Ref
4  ◌  2 ── %2  = $(Expr(:enter, #8))                                           │
5  ◌  3 ── %3  = Base.isdefined(%1, :x)::Bool                                  │╻╷ get′
   ◌  └───       goto #5 if not %3                                             ││
   ↑  4 ── %5  = Base.getfield(%1, :x)::String                                 ││╻  getindex
   ◌  └───       goto #6                                                       ││
   ◌  5 ──       Main.throw(%1)::Union{}                                       ││
   ◌  └───       unreachable                                                   ││
   ◌  6 ──       $(Expr(:leave, 1))                                            │
   ◌  7 ──       goto #10                                                      │
   ◌  8 ──       $(Expr(:leave, 1))                                            │
   ◌  9 ── %12 = $(Expr(:the_exception))::Any                                  │
7  ↑  │    %13 = Main.typeof(%12)::DataType                                    │
8  ◌  │          invoke Main.rethrow_escape!()::Any                            │
   ◌  └───       $(Expr(:pop_exception, :(%2)))::Any                           │
10 ↑  10 ┄ %16 = φ (#7 => %5, #9 => %13)::Union{DataType, String}              │
   ◌  └───       return %16                                                    │

It requires a global analysis in order to correctly reason about all possible escapes via existing exception interfaces. For now we always propagate the topmost escape information to all potentially thrown objects conservatively, since such an additional analysis might not be worthwhile to do given that exception handling and error path usually don't need to be very performance sensitive, and also optimizations of error paths might be very ineffective anyway since they are often even "unoptimized" intentionally for latency reasons.

x::EscapeInfo's x.ThrownEscape property records SSA statements where x can be thrown as an exception. Using this information EscapeAnalysis can propagate possible escapes via exceptions limitedly to only those may be thrown in each try region:

julia> result = code_escapes((String,String)) do s1, s2
           r1 = Ref(s1)
           r2 = Ref(s2)
           local ret
           try
               s1 = get′(r1)
               ret = sizeof(s1)
           catch err
               global g = err # will definitely escape `r1`
           end
           s2 = get′(r2)      # still `r2` doesn't escape fully
           return s2
       end
#11(X _2::String, ↑ _3::String) in Main at REPL[13]:2
2  X  1 ── %1  = %new(Base.RefValue{String}, _2)::Base.RefValue{String}   │╻╷╷ Ref
3  *′ └─── %2  = %new(Base.RefValue{String}, _3)::Base.RefValue{String}   │╻╷╷ Ref
5  ◌  2 ── %3  = $(Expr(:enter, #8))                                      │
   *′ └─── %4  = ϒ (%2)::Base.RefValue{String}                            │
6  ◌  3 ── %5  = Base.isdefined(%1, :x)::Bool                             │╻╷  get′
   ◌  └───       goto #5 if not %5                                        ││
   X  4 ──       Base.getfield(%1, :x)::String                            ││╻   getindex
   ◌  └───       goto #6                                                  ││
   ◌  5 ──       Main.throw(%1)::Union{}                                  ││
   ◌  └───       unreachable                                              ││
7  ◌  6 ──       nothing::typeof(Core.sizeof)                             │╻   sizeof
   ◌  │          nothing::Int64                                           ││
   ◌  └───       $(Expr(:leave, 1))                                       │
   ◌  7 ──       goto #10                                                 │
   *′ 8 ── %15 = φᶜ (%4)::Base.RefValue{String}                           │
   ◌  └───       $(Expr(:leave, 1))                                       │
   X  9 ── %17 = $(Expr(:the_exception))::Any                             │
9  ◌  │          (Main.g = %17)::Any                                      │
   ◌  └───       $(Expr(:pop_exception, :(%3)))::Any                      │
11 *′ 10 ┄ %20 = φ (#7 => %2, #9 => %15)::Base.RefValue{String}           │
   ◌  │    %21 = Base.isdefined(%20, :x)::Bool                            ││╻   isassigned
   ◌  └───       goto #12 if not %21                                      ││
   ↑  11 ─ %23 = Base.getfield(%20, :x)::String                           │││╻   getproperty
   ◌  └───       goto #13                                                 ││
   ◌  12 ─       Main.throw(%20)::Union{}                                 ││
   ◌  └───       unreachable                                              ││
12 ◌  13 ─       return %23                                               │

Analysis Usage

When using EscapeAnalysis in Julia's high-level compilation pipeline, we can run analyze_escapes(ir::IRCode) -> estate::EscapeState to analyze escape information of each SSA-IR element in ir.

Note that it should be most effective if analyze_escapes runs after inlining, as EscapeAnalysis's interprocedural escape information handling is limited at this moment.

Since the computational cost of analyze_escapes is not that cheap, it is more ideal if it runs once and succeeding optimization passes incrementally update the escape information upon IR transformation.

EscapeAnalysis.analyze_escapes — Function

analyze_escapes(ir::IRCode, nargs::Int) -> estate::EscapeState

Analyzes escape information in ir. nargs is the number of actual arguments of the analyzed call.

EscapeAnalysis.EscapeState — Type

estate::EscapeState

Extended lattice that maps arguments and SSA values to escape information represented as EscapeInfo. Escape information imposed on SSA IR element x can be retrieved by estate[x].

EscapeAnalysis.EscapeInfo — Type

x::EscapeInfo

A lattice for escape information, which holds the following properties:

x.Analyzed::Bool: not formally part of the lattice, only indicates x has not been analyzed or not

x.ReturnEscape::Bool: indicates x can escape to the caller via return

x.ThrownEscape::BitSet: records SSA statements numbers where x can be thrown as exception: this information will be used by escape_exception! to propagate potential escapes via exception

x.AliasInfo::Union{IndexableFields,Unindexable,Bool}: maintains all possible values that can be aliased to fields or array elements of x:

x.AliasInfo === false indicates the fields/elements of x isn't analyzed yet

x.AliasInfo === true indicates the fields/elements of x can't be analyzed, e.g. the type of x is not known or is not concrete and thus its fields/elements can't be known precisely

x.AliasInfo::IndexableFields records all the possible values that can be aliased to fields of object x with precise index information

x.AliasInfo::IndexableElements records all the possible values that can be aliased to elements of array x with precise index information

x.AliasInfo::Unindexable records all the possible values that can be aliased to fields/elements of x without precise index information

x.Liveness::BitSet: records SSA statement numbers where x should be live, e.g. to be used as a call argument, to be returned to a caller, or preserved for :foreigncall. 0 ∈ x.Liveness has the special meaning that it's a call argument of the currently analyzed call frame (and thus it's visible from the caller immediately).

x.ArgEscape::Int (not implemented yet): indicates it will escape to the caller through setfield! on argument(s)

-1 : no escape

0 : unknown or multiple

n : through argument N

There are utility constructors to create common EscapeInfos, e.g.,

NoEscape(): the bottom(-like) element of this lattice, meaning it won't escape to anywhere

AllEscape(): the topmost element of this lattice, meaning it will escape to everywhere

analyze_escapes will transition these elements from the bottom to the top, in the same direction as Julia's native type inference routine. An abstract state will be initialized with the bottom(-like) elements:

the call arguments are initialized as ArgEscape(), whose Liveness property includes 0 to indicate that it is passed as a call argument and visible from a caller immediately

the other states are initialized as NotAnalyzed(), which is a special lattice element that is slightly lower than NoEscape, but at the same time doesn't represent any meaning other than it's not analyzed yet (thus it's not formally part of the lattice)

EscapeAnalysis.cache_escapes! — Function

cache_escapes!(linfo::MethodInstance, estate::EscapeState, _::IRCode)

Transforms escape information of estate for interprocedural propagation, and caches it in a global cache that can then be looked up later when linfo callsite is seen again.

LatticeDesignOur type inference implementation takes the alternative approach, where each lattice property is represented by a special lattice element type object. It turns out that it started to complicate implementations of the lattice operations mainly because it often requires conversion rules between each lattice element type object. And we are working on overhauling our type inference lattice implementation with EscapeInfo-like lattice design.

MM02A Graph-Free approach to Data-Flow Analysis. Markas Mohnen, 2002, April. https://api.semanticscholar.org/CorpusID:28519618.

BackandForthOur type inference algorithm in contrast is implemented as a forward analysis, because type information usually flows from "definition" to "usage" and it is more natural and effective to propagate such information in a forward way.

DynamismIn some cases, however, object fields can't be analyzed precisely. For example, object may escape to somewhere EscapeAnalysis can't account for possible memory effects on it, or fields of the objects simply can't be known because of the lack of type information. In such cases AliasInfo property is raised to the topmost element within its own lattice order, and it causes succeeding field analysis to be conservative and escape information imposed on fields of an unanalyzable object to be propagated to the object itself.

JVM05Escape Analysis in the Context of Dynamic Compilation and Deoptimization. Thomas Kotzmann and Hanspeter Mössenböck, 2005, June. https://dl.acm.org/doi/10.1145/1064979.1064996.